Changed pressure to a float.

+++ elements.h
@@ -34,7 +34,7 @@ public:
   std::string name; 
   bool fixed;
   float density;
-  int pressure;
+  float pressure;
   unsigned int argb;
   ProbList decay;
   ReactionList reactions;
